Rename test script
